Split CourseModal
